requirements.txt
requests==2.28.1
environs~=9.5.0
websockets==10.4
sseclient-py==1.7.2
urllib3~=1.26.12
gpiozero~=1.6.2
qrcode~=7.3.1
Pillow~=9.3.0